fix: shorten header filenames to fix Windows MAX_PATH install failure
Some installed headers under filter/scalar_topology/worklet/ had relative
paths exceeding 130 characters. Combined with a typical conda package cache
prefix (~104 chars), these paths exceeded the Windows MAX_PATH limit of 260
characters, making installation impossible without Long Path support.

Rename 14 offending headers to shorter equivalents and update all include
directives, include guards, and CMakeLists.txt entries accordingly.

Also add Utilities/Scripts/check-path-lengths.bash and a matching CI job
to enforce a 130-character limit on all installed headers going forward.

Fixes #318
