+++ viskores/filter/scalar_topology/worklet/contourtree_distributed/HierarchicalHyperSweeper.h
@@ -94,9 +94,9 @@
 #include <viskores/filter/scalar_topology/worklet/contourtree_distributed/PrintGraph.h>
 #include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/ComputeSuperarcDependentWeightsWorklet.h>
 #include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/ComputeSuperarcTransferWeightsWorklet.h>
-#include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/InitializeIntrinsicVertexCountComputeSuperparentIdsWorklet.h>
-#include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/InitializeIntrinsicVertexCountInitalizeCountsWorklet.h>
-#include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/InitializeIntrinsicVertexCountSubtractLowEndWorklet.h>
+#include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/InitIntrinsicVertexCountInitCountsWorklet.h>
+#include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/InitIntrinsicVertexCountSubtractLowEndWorklet.h>
+#include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/InitIntrinsicVertexCountSuperparentIds.h>
 #include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/TransferTargetComperator.h>
 #include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/TransferWeightsUpdateLHEWorklet.h>
 #include <viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_hyper_sweeper/TransferWeightsUpdateLHEWorkletRound2.h>
