+++ viskores/filter/scalar_topology/worklet/contourtree_distributed/hierarchical_augmenter/CMakeLists.txt
@@ -33,7 +33,7 @@ set(headers
   FillEmptyIterationWorklet.h
   CreateSuperarcsWorklet.h
   CreateSuperarcsData.h
-  CreateSuperarcsSetFirstSupernodePerIterationWorklet.h
+  CreateSuperarcsSetFirstSupernodePerIterWorklet.h
   HierarchicalAugmenterInOutData.h
 )
